aoc_2025/pyproject.toml
[project]
name = "aoc-2025"
version = "0.1.0"
description = ""
authors = [
    {name = "Krzysztof Jurkiewicz",email = "krzysztof.jurkiewicz@zowie.ai"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["util"]
